requirements.txt
Flask==2.3.3
msal==1.21.0
oauthlib==3.2.2
requests==2.28.2
python-dotenv==1.0.0
gunicorn==23.0.0